+++ package.json
@@ -47,6 +47,7 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.11.1",
+    "ignore-styles": "^5.0.1",
     "istanbul": "^0.4.5",
     "jsdom": "^13.0.0",
     "mocha": "^5.2.0",
