+++ test/lib/mocha.opts
@@ -1,2 +1,3 @@
 --require @babel/register
+--require ignore-styles
 --file ./test/lib/mocha.setup.js
